+++ apps/ehr/src/constants/feature-flags.ts
@@ -1,4 +1,4 @@
 export const FEATURE_FLAGS = {
   LAB_ORDERS_ENABLED: import.meta.env.VITE_APP_IS_LAB_ORDERS_ENABLED_FEATURE_FLAG === 'true',
-  IN_HOUSE_LABS_ENABLED: import.meta.env.VITE_APP_IS_IN_HOUSE_LABS_ENABLED_ENABLED_FEATURE_FLAG === 'true',
+  IN_HOUSE_LABS_ENABLED: import.meta.env.VITE_APP_IS_IN_HOUSE_LABS_ENABLED_FEATURE_FLAG === 'true',
 };
