+++ apps/ehr/src/features/in-house-labs/components/details/FinalResultView.tsx
