+++ apps/ehr/src/features/in-house-labs/components/details/PerformTestView.tsx
@@ -10,13 +10,11 @@ import {
   RadioGroup,
   Checkbox,
   Collapse,
-  TextField,
-  Divider,
 } from '@mui/material';
 import KeyboardArrowDownIcon from '@mui/icons-material/KeyboardArrowDown';
 import KeyboardArrowUpIcon from '@mui/icons-material/KeyboardArrowUp';
 import { LabTest, TestResult } from '../../labTypes';
-import { DateTime } from 'luxon';
+import { History } from './History';
 
 interface PerformTestViewProps {
   testDetails: LabTest;
@@ -36,14 +34,6 @@ export const PerformTestView: React.FC<PerformTestViewProps> = ({ testDetails, o
 
   const handleReprintLabel = (): void => {
     console.log('Reprinting label for test:', testDetails.id);
-  };
-
-  const formatDateTime = (dateString: string): string => {
-    try {
-      return DateTime.fromISO(dateString).toFormat("MM/dd/yyyy 'at' h:mm a");
-    } catch (error) {
-      return dateString;
-    }
   };
 
   const handleSubmit = (): void => {
@@ -179,78 +169,12 @@ export const PerformTestView: React.FC<PerformTestViewProps> = ({ testDetails, o
           </Box>
 
           <Collapse in={showDetails}>
-            <Box mb={3}>
-              <Box sx={{ mt: 3, mb: 1 }}>
-                <Typography variant="body2" sx={{ mb: 1 }}>
-                  Provider notes
-                </Typography>
-                <TextField fullWidth multiline rows={4} value={notes} onChange={(e) => setNotes(e.target.value)} />
-              </Box>
-
-              {testDetails.orderDetails && (
-                <Box mt={3}>
-                  <Button variant="outlined" onClick={handleReprintLabel} sx={{ mb: 3 }}>
-                    Re-Print Label
-                  </Button>
-
-                  <Grid container sx={{ backgroundColor: '#F8F9FA', p: 2 }}>
-                    <Grid item xs={3}>
-                      <Box
-                        sx={{
-                          bgcolor: '#F1F3F4',
-                          color: '#5F6368',
-                          fontWeight: 'bold',
-                          px: 2,
-                          py: 0.5,
-                          borderRadius: '4px',
-                          fontSize: '0.75rem',
-                          display: 'inline-block',
-                        }}
-                      >
-                        ORDERED
-                      </Box>
-                    </Grid>
-                    <Grid item xs={4}>
-                      <Typography variant="body2">{testDetails.orderDetails.orderedBy}</Typography>
-                    </Grid>
-                    <Grid item xs={5}>
-                      <Typography variant="body2">{formatDateTime(testDetails.orderDetails.orderedDate)}</Typography>
-                    </Grid>
-
-                    <Grid item xs={12} sx={{ mt: 1 }}>
-                      <Divider />
-                    </Grid>
-
-                    <Grid item xs={3} sx={{ mt: 1 }}>
-                      <Box
-                        sx={{
-                          bgcolor: '#E8DEFF',
-                          color: '#5E35B1',
-                          fontWeight: 'bold',
-                          px: 2,
-                          py: 0.5,
-                          borderRadius: '4px',
-                          fontSize: '0.75rem',
-                          display: 'inline-block',
-                        }}
-                      >
-                        COLLECTED
-                      </Box>
-                    </Grid>
-                    <Grid item xs={4} sx={{ mt: 1 }}>
-                      <Typography variant="body2">{testDetails.orderDetails.collectedBy || ''}</Typography>
-                    </Grid>
-                    <Grid item xs={5} sx={{ mt: 1 }}>
-                      <Typography variant="body2">
-                        {testDetails.orderDetails.collectedDate
-                          ? formatDateTime(testDetails.orderDetails.collectedDate)
-                          : ''}
-                      </Typography>
-                    </Grid>
-                  </Grid>
-                </Box>
-              )}
-            </Box>
+            <History
+              testDetails={testDetails}
+              setNotes={setNotes}
+              notes={notes}
+              handleReprintLabel={handleReprintLabel}
+            />
           </Collapse>
 
           <Box display="flex" justifyContent="space-between">
