+++ apps/ehr/src/features/in-house-labs/labConstants.ts
@@ -1,0 +1,7 @@
+import { TestStatus } from './labTypes';
+
+export const inHouseLabsTestStatuses: Record<TestStatus, TestStatus> = {
+  ORDERED: 'ORDERED',
+  COLLECTED: 'COLLECTED',
+  FINAL: 'FINAL',
+};
