+++ apps/ehr/src/features/in-house-labs/labTypes.ts
@@ -1,10 +1,10 @@
 // Types of lab tests
-export type TestType = 'QUALITATIVE' | 'QUANTITATIVE' | 'URINALYSIS';
+export type TestType = 'QUALITATIVE' | 'QUANTITATIVE' | 'MIXED';
 
 // Possible test statuses
 export type TestStatus = 'ORDERED' | 'COLLECTED' | 'FINAL';
 
-// Types of test results
+// todo: it's a draft, see a comment https://github.com/masslight/ottehr/pull/2166#discussion_r2085316003
 export type TestResult = 'DETECTED' | 'NOT_DETECTED' | 'INDETERMINATE' | null;
 
 // Urine analysis parameter type
